+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update `chat` example to webdev.
 - Allow nulls when serializing/deserializing for better JSON interop.
 - Fix generation bugs around enum wire name and polymorphism.
+- Fix generation with generics for analysis with `strict-raw-types`.
 
 # 8.0.0